requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/dimka665/awesome-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise[brotli]==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.4  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-redis==5.2.0  # https://github.com/niwinz/django-redis
# NOTE: there's an issue with 1.14 and bootstrap4 forms not having form-control class
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
# DRF-spectacular for api documentation
drf-spectacular==0.26.2  # https://github.com/tfranzel/drf-spectacular
djangorestframework-csv==2.1.1

# MY STUFF
django-recaptcha==3.0.0  # https://github.com/praekelt/django-recaptcha
sentry-sdk==1.21.1  # https://github.com/getsentry/sentry-python/

metapub==0.5.5
biopython==1.81
pandas==2.0.1
parasail==1.3.4

habanero==1.2.3
django-filter==23.2
django-reversion==5.0.4
django-reversion-compare==0.16.1

tablib==3.4.0
scipy==1.10.1
dnspython==2.3.0  #
django-anymail==9.1  # https://github.com/anymail/django-anymail
django-autocomplete-light==3.9.4  # https://github.com/yourlabs/django-autocomplete-light
django-avatar==7.1.1

numpy==1.24.3
matplotlib==3.7.1
django-mptt==0.14.0  # https://github.com/django-mptt/django-mptt
django-webpack-loader==1.8.1

google-api-python-client==2.86.0
oauth2client==4.1.3
algoliasearch_django==3.0.0

celery[redis]==5.2.7  # https://github.com/celery/celery
graphene==3.2.2  # https://github.com/graphql-python/graphene
graphene-django==3.0.2  # https://github.com/graphql-python/graphene-django
graphene-django-optimizer==0.9.1  # https://github.com/tfoxy/graphene-django-optimizer
django-graphql-jwt==0.3.4
django-cors-headers==3.14.0
